topographic/requirements.txt
numpy==1.18.1
nilearn.egg==info
seaborn==0.10.0
easydict==1.9
imageio==2.8.0
torch==1.4.0
torchvision==0.5.0
pytorch_nlp==0.5.0
scipy==1.5.3
nibabel==2.3.1
torchsummary==1.5.1
statsmodels==0.11.1
apex==0.1
psutil==5.7.0
tqdm==4.42.1
pandas==1.0.5
cmasher==1.3.5
scikit_image==0.16.2
matplotlib==3.1.3
nilearn==0.8.0
Pillow==8.3.1
scikit_learn==0.24.2
skimage==0.0
statannot==0.2.3
tensorboardX==2.4
torchnlp==0.0.0.1
